README.md: Correct out of data information
First off, thanks for this work! I just started reading SCP and was interested in if there was any public API for it. Finding this was exciting! I decided, just for fun, to make an API client for your site / repo. While working on the deserializer, I found a few bits of your API documentation that were incorrect. There may be more, I don't know. I'm not finished with it :)